+++ js/checkin.js
@@ -1,0 +1,3 @@
+const checkinArray =  ['12:00', '13:00', '14:00'];
+
+export {checkinArray};
