+++ js/checkout.js
@@ -1,0 +1,1 @@
+export const checkoutArray = ['12:00', '13:00', '14:00'];
