+++ js/typesArray.js
@@ -1,0 +1,4 @@
+
+const typesArray = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+
+export {typesArray};
